billing_api/pyproject.toml
[tool.poetry]
name = "billing-api"
version = "0.1.0"
description = "Billing API service"
authors = ["Yandex Practicum Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
pydantic-settings = "^2.10.1"
sqlalchemy = "^2.0.43"
gunicorn = "^23.0.0"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.0"
python-jose = "3.5.0"
aiokafka = "^0.10.0"
httpx = "^0.28.1"
alembic = "1.13.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

